+++ src/errors.ts
@@ -1,13 +1,26 @@
-/**
- * A simple implementation of a command line interface definer
- */
+export enum ParserExitCodes {
+  NO_ERROR,
+  RESERVED = 1000,
+  HELP_AND_EXIT, // the parser says exit, but nothing's wrong
+  JUST_EXIT, // exit with a zero code, don't run the command
+  UNRECOGNIZED_SUBCOMMAND,
+  UNRECOGNIZED_FLAG,
+  MISSING_REQUIRED_FLAG,
+  INVALID_FLAG_ARGS,
+  UNKNOWN_ERROR = 1500,
+}
 
 /**
  * ParsingError gives us a way to throw errors
  * that might happen during parsing with extra metadata to
  */
 export class ParsingError extends Error {
-  constructor(message: string, private advice = "", private paramName = "") {
+  constructor(
+    message: string,
+    readonly code: ParserExitCodes,
+    readonly advice = "",
+    readonly paramName = "",
+  ) {
     super(message);
   }